Removing unused code, and renaming some utilities.
